LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/getClientRects-inline-with-block-child-expected.txt


FAIL getClientRects-inline-with-block-child assert_greater_than_equal: At least 2 fragments expected a number greater than or equal to 2 but got 1
